pyproject.toml
[tool.poetry]
name = "financial_timeseries_pipeline"
version = "0.1.0"
description = "Financial Time Series Data Ingestion Pipeline"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.0"
yfinance = "^0.2.36"
alpha_vantage = "^2.3.1"
influxdb-client = "^1.36.0"
python-dotenv = "^1.0.0"
apache-airflow = "^2.7.0"
pydantic = "^2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
black = "^22.0.0"
flake8 = "^4.0.0"
mypy = "^1.0.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"